kwi: saido ba o seri simasita
