+++ src/Qiwngjyong.vue
@@ -6,13 +6,9 @@ import KriemsakTyolan from '@/cogrien/KriemsakTyolan.vue'
 import KriemsakTyolanChrioshiw from './cogrien/KriemsakTyolanChrioshiw.vue'
 import { jyongXeithongkho } from '@/changkho/xeithong'
 import { jyongDziwdeukho } from '@/changkho/dziwdeu'
-import { useTheme as jyongTyodei } from 'vuetify'
 
 const xeithongkho = jyongXeithongkho()
 const dziwdeukho = jyongDziwdeukho()
-const tyodei = jyongTyodei()
-
-const xuanTyodei = () => (tyodei.global.name.value = tyodei.global.current.value.dark ? 'light' : 'dark')
 </script>
 
 <template>
@@ -31,13 +27,8 @@ const xuanTyodei = () => (tyodei.global.name.value = tyodei.global.current.value
       <v-app-bar-nav-icon @click="xeithongkho.xuanKriemsakPenlanThryutheiCatia()" />
       <v-app-bar-title>漢字文聲義</v-app-bar-title>
       <v-btn
-        @click="xuanTyodei()"
-        icon="mdi-lightbulb"
-      >
-      </v-btn>
-      <v-btn
         @click="xeithongkho.xuanKriemsakPenlanThryutheiXyutia()"
-        icon="mdi-bookmark-multiple-outline"
+        icon="mdi-cog-outline"
       >
       </v-btn>
     </v-app-bar>
