+++ src/cisren/tyoqieu.scss
@@ -1,6 +1,6 @@
 .qhwkthei {
   font-family: 'AndikaW', '思源黑体 CN VF', 'Source Han Sans CN VF', '思源黑体 CN', 'Source Han Sans CN', '思源黑体 VF',
-    'Source Han Sans SC VF', '思源黑体', 'Source Han Sans SC', 'Noto Sans SC', sans-serif;
+    'Source Han Sans SC VF', '思源黑体', 'Source Han Sans SC', 'Noto Sans SC', 'PingFang SC', sans-serif;
 }
 
 .songthei {
