+++ src/cogrien/KriemsakPenlanCatia.vue
@@ -1,21 +1,10 @@
 <script setup lang="ts">
 import { jyongXeithongkho } from '@/changkho/xeithong'
-import { jyongDziwbyoKriemsakkho } from '@/changkho/dziwbyoKriemsak'
-import { jyongDziwqrimKriemsakkho } from '@/changkho/dziwqrimKriemsak'
-import { jyongDziwdeukho } from '@/changkho/dziwdeu'
-import shiwxryn from '../sryokio/shiwxryn.json'
-import shiengmu from '../sryokio/shiengmu.json'
+import KriemsakPenlanCatiaKriemsak from './KriemsakPenlanCatiaKriemsak.vue'
+import KriemsakPenlanCatiaShiochiem from './KriemsakPenlanCatiaShiochiem.vue'
+import KriemsakPenlanCatiaLeksriw from './KriemsakPenlanCatiaLeksriw.vue'
 
 const xeithongkho = jyongXeithongkho()
-const dziwbyoKriemsakkho = jyongDziwbyoKriemsakkho()
-const dziwqrimKriemsakkho = jyongDziwqrimKriemsakkho()
-const dziwdeukho = jyongDziwdeukho()
-
-const qanKriemsakKetkua = (tuiziang: string) => {
-  dziwdeukho.triwSrioSyenDziwdu(tuiziang)
-}
-
-const qhruaSikrokXaumra = (xaumraco: string[]) => xaumraco.join('、')
 </script>
 
 <template>
@@ -24,100 +13,20 @@ const qhruaSikrokXaumra = (xaumraco: string[]) => xaumraco.join('、')
     location="left"
   >
     <v-tabs
-      v-model="xeithongkho.kriemsakPenlanCatiaPieuchiem"
+      v-model="xeithongkho.kriemsakPenlanCatiaPieuchiemDeiqit"
       fixed-tabs
     >
-      <v-tab value="krap">字符</v-tab>
-      <v-tab value="qrit">字音</v-tab>
+      <v-tab value="krap"><v-icon> mdi-magnify </v-icon></v-tab>
+      <v-tab value="qrit"><v-icon> mdi-bookmark-multiple-outline </v-icon></v-tab>
+      <v-tab value="prieng"><v-icon> mdi-history </v-icon></v-tab>
     </v-tabs>
-    <v-window v-model="xeithongkho.kriemsakPenlanCatiaPieuchiem">
+    <v-window v-model="xeithongkho.kriemsakPenlanCatiaPieuchiemDeiqit">
       <v-window-item value="krap">
-        <v-sheet class="pa-4">
-          <v-text-field
-            v-model="dziwbyoKriemsakkho.srioNipQhandziwthyen"
-            label="漢字"
-            placeholder="天地玄黄宇宙洪荒"
-            clearable
-            @keydown.enter="dziwbyoKriemsakkho.triwKriemsakKetkua()"
-          />
-          <v-btn
-            block
-            rounded
-            @click="dziwbyoKriemsakkho.triwKriemsakKetkua()"
-          >
-            檢索
-          </v-btn>
-        </v-sheet>
-
-        <v-divider />
-
-        <v-list>
-          <v-list-item
-            v-for="tuiziang in dziwbyoKriemsakkho.twkKriemsakKetkua"
-            :key="tuiziang.pieutiwbyo"
-            @click="qanKriemsakKetkua(tuiziang.dziw)"
-            link
-          >
-            <v-list-item-title>{{ tuiziang.dziw }}</v-list-item-title>
-            {{ qhruaSikrokXaumra(tuiziang.sikrok_xaumra) }}
-          </v-list-item>
-        </v-list>
+        <KriemsakPenlanCatiaKriemsak />
       </v-window-item>
-      <v-window-item value="qrit">
-        <v-sheet class="pa-4">
-          <v-col>
-            <v-text-field
-              v-model="dziwqrimKriemsakkho.srioNipZyepheng"
-              label="隋拼"
-              placeholder="tuŋ"
-              density="compact"
-              variant="underlined"
-              clearable
-              @keydown.enter="dziwqrimKriemsakkho.triwKriemsakKetkua"
-            />
-            <v-select
-              label="平水韵"
-              v-model="dziwqrimKriemsakkho.srioSyenShiwxryn"
-              :items="shiwxryn"
-              item-title="xryn_driang"
-              item-value="xryn_tuan"
-              variant="underlined"
-              density="compact"
-              clearable
-            />
-            <v-select
-              label="聲母"
-              v-model="dziwqrimKriemsakkho.srioSyenShiengmu"
-              :items="shiengmu"
-              item-title="shiengmu_ziangdzieng"
-              item-value="shiengmu"
-              variant="underlined"
-              density="compact"
-              clearable
-            />
-          </v-col>
-          <v-btn
-            block
-            rounded
-            @click="dziwqrimKriemsakkho.triwKriemsakKetkua"
-          >
-            檢索
-          </v-btn>
-        </v-sheet>
-
-        <v-divider />
-
-        <v-list>
-          <v-list-item
-            v-for="tuiziang in dziwqrimKriemsakkho.twkKriemsakKetkua"
-            :key="tuiziang.pieutiwbyo"
-            @click="qanKriemsakKetkua(tuiziang.dziw)"
-            link
-          >
-            <v-list-item-title>{{ tuiziang.dziw }}</v-list-item-title>
-            {{ tuiziang.zyepheng }}
-          </v-list-item>
-        </v-list>
+      <v-window-item value="qrit"> <KriemsakPenlanCatiaShiochiem /> </v-window-item>
+      <v-window-item value="prieng">
+        <KriemsakPenlanCatiaLeksriw />
       </v-window-item>
     </v-window>
   </v-navigation-drawer>
